+++ python/katana/local/analytics/_ksssp.pyx
@@ -69,8 +69,8 @@ cdef extern from "katana/analytics/k_shortest_paths/ksssp.h" namespace "katana::
         _AlgoReachability SyncLevel()
 
     Result[void] Ksssp(_PropertyGraph* pg, const string& edge_weight_property_name, 
-                       uint32_t start_node, uint32_t report_node, CTxnContext* txn_ctx, 
-                       _AlgoReachability algo_reachability, uint32_t num_paths, 
+                       size_t start_node, size_t report_node, CTxnContext* txn_ctx, 
+                       _AlgoReachability algo_reachability, size_t num_paths, 
                        const bool& is_symmetric, _KssspPlan plan)
 
 class _KssspAlgorithm(Enum):
@@ -226,8 +226,8 @@ cdef class AlgoReachability:
         return AlgoReachability.make(_AlgoReachability.SyncLevel())
 
     
-def ksssp(pg, str edge_weight_property_name, int start_node, 
-          int report_node, int num_paths, bool is_symmetric=False, 
+def ksssp(pg, str edge_weight_property_name, size_t start_node, 
+          size_t report_node, size_t num_paths, bool is_symmetric=False, 
           AlgoReachability algo_reachability = AlgoReachability(), 
           KssspPlan plan = KssspPlan(), *, txn_ctx = None):
     """
